requirements.txt
fastapi==0.115.12
google-generativeai==0.8.5
pydantic==2.11.5
uvicorn===0.34.2
python-dotenv==1.1.0